requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
beautifulsoup4==4.12.3
discord.py==2.3.2
feedparser==6.0.11
lxml==5.2.1
Requests==2.31.0
